# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/accounts": {
            "get": {
                "description": "List Accounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Banking",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.AccountList"
                        }
                    }
                }
            },
            "post": {
                "description": "create bank account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Banking",
                "parameters": [
                    {
                        "description": "Create Account",
                        "name": "AccountInputSchema",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.AccountInputSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.AccountOutputSchema"
                        }
                    }
                }
            }
        },
        "/api/v1/accounts/{id}": {
            "get": {
                "description": "get account by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Banking",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Find Account by Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.AccountOutputSchema"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "rest.AccountInputSchema": {
            "type": "object",
            "properties": {
                "document": {
                    "type": "string"
                }
            }
        },
        "rest.AccountList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.AccountOutputSchema"
                    }
                }
            }
        },
        "rest.AccountOutputSchema": {
            "type": "object",
            "properties": {
                "document": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}